Split up rendering and comparison logic
